refactor: organize package, move 'extensions.Cache' to 'cache.Extension'
